Phase 4: MediaPipeline owns assembly, the pump goes back to being transport-only
session::pipeline picks the sinks, wires the stages and owns the three threads
plus one shutdown ordering. connect.rs is the handshake again.

AU reassembly moves down with it: the pump hands over a WireFrame and the stage
decides what it means - which pieces fit together, whether the timeline holds,
what the decoder's clock calls this frame. The pump no longer knows AUs arrive in
pieces at all.
