+++ src/session/pump.rs
@@ -1,5 +1,6 @@
-//! The threads that drain the transport: video into the sink, audio into NDL or the
-//! software playback ring.
+//! The threads that drain the transport into the pipeline: access units into the video stage,
+//! packets into the audio stage. Everything wire-shaped lives here and nothing else — what a
+//! delivery MEANS is the stages' business.
 
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
@@ -14,7 +15,7 @@ use crate::platform::webos::device::boost_current_thread;
 use crate::session::audio::AudioStage;
 use crate::session::join::{join_with_timeout, SHUTDOWN_JOIN_TIMEOUT};
 use crate::session::priority::{boost_hot_threads, spawn_vendor_decode_thread_renicer};
-use crate::session::stage::{FrameFlags, SinkResult, VideoStage};
+use crate::session::stage::{SinkResult, VideoStage, WireFrame};
 use crate::session::StreamStats;
 
 /// Longest a `next_frame` call parks before the loop re-checks `stop`.
@@ -50,92 +51,6 @@ impl Tick {
     }
 }
 
-/// What [`AuParts`] decided about one delivery.
-enum PartStep {
-    /// Hand the bytes to the sink. `partial` = this is not the AU's last piece; `lost_parts` = an
-    /// earlier AU died mid-flight, so decoding cannot continue from where it stopped.
-    Feed { partial: bool, lost_parts: bool },
-    /// Nothing usable — a piece of an AU already abandoned. The decoder must not see it.
-    Discard,
-}
-
-/// Slice-progressive reassembly bookkeeping (`punktfunk_core::session::FramePart`).
-///
-/// On every NDL v2 session, AU prefixes arrive while the rest is still on the wire and the decoder
-/// gets a frame's first bytes without waiting for its last datagram — a real slice of a frame
-/// period at high bitrate, and pure latency: none of that wait is decode work. On a backend that
-/// can't take them (`Negotiated::clamp`) every delivery carries `part: None` and this is a
-/// pass-through.
-///
-/// The contract enforced here is core's: parts arrive in order with no gaps, BUT the pre-decode
-/// hand-off may drop entries (memory pressure, a jump-to-live clear), so an `offset` that isn't the
-/// open AU's next expected byte means that AU is gone. There is no abort marker — a `first` part for
-/// a new index while one is still open is how a death is signalled. Both cases abandon the AU and
-/// report loss, which puts the sink into freeze-until-reanchor and asks the host for a keyframe:
-/// the decoder is holding a truncated input, and nothing short of a fresh anchor clears it.
-#[derive(Default)]
-struct AuParts {
-    /// `(frame_index, next expected byte offset)` of the AU currently being fed.
-    open: Option<(u32, u32)>,
-    /// An AU was abandoned and nothing has restarted decoding since — so whatever comes next is
-    /// resuming against a decoder that still holds a truncated frame.
-    abandoned: bool,
-}
-
-impl AuParts {
-    fn step(&mut self, frame: &punktfunk_core::session::Frame, takes_parts: bool) -> PartStep {
-        let Some(part) = frame.part.filter(|_| takes_parts) else {
-            // Whole-AU delivery: parts weren't negotiated, this backend doesn't take them, or
-            // this is an aged-out chunk-aligned partial — core hands all three over as one buffer.
-            self.open = None;
-            return PartStep::Feed {
-                partial: false,
-                lost_parts: std::mem::take(&mut self.abandoned),
-            };
-        };
-        let len = frame.data.len() as u32;
-        if part.first {
-            let lost_parts = self.open.take().is_some() | std::mem::take(&mut self.abandoned);
-            if lost_parts {
-                tracing::warn!("frame parts: AU {} starts over an unfinished one", frame.frame_index);
-            }
-            self.open = (!part.last).then_some((frame.frame_index, len));
-            return PartStep::Feed {
-                partial: !part.last,
-                lost_parts,
-            };
-        }
-        match self.open {
-            Some((index, next)) if index == frame.frame_index && next == part.offset => {
-                self.open = (!part.last).then_some((index, next + len));
-                PartStep::Feed {
-                    partial: !part.last,
-                    lost_parts: false,
-                }
-            }
-            // Either nothing is open (the AU was abandoned, or this part arrived without its head)
-            // or the offset skipped — both mean the AU can never be completed.
-            open => {
-                if open.is_some() {
-                    tracing::warn!(
-                        "frame parts: AU {} broke at offset {} — abandoning",
-                        frame.frame_index,
-                        part.offset,
-                    );
-                }
-                self.drop_open();
-                PartStep::Discard
-            }
-        }
-    }
-
-    /// Forget the AU in flight: its remaining parts are no longer feedable, and whatever restarts
-    /// decoding has to be told the decoder holds a truncated input.
-    fn drop_open(&mut self) {
-        self.abandoned |= self.open.take().is_some();
-    }
-}
-
 /// Drives the video thread: transport → [`VideoStage`], plus the counters and the loss/HDR
 /// side-channels that ride the same loop.
 struct VideoPump {
@@ -149,9 +64,6 @@ struct VideoPump {
     last_dropped_seen: u64,
     heartbeat: Tick,
     video_log: Tick,
-    /// Slice-progressive reassembly state — a pass-through on a backend that doesn't take parts
-    /// (see [`AuParts`]).
-    parts: AuParts,
 }
 
 impl VideoPump {
@@ -165,7 +77,6 @@ impl VideoPump {
             last_dropped_seen,
             heartbeat: Tick::new(HEARTBEAT),
             video_log: Tick::new(VIDEO_LOG_INTERVAL),
-            parts: AuParts::default(),
         }
     }
 
@@ -205,38 +116,24 @@ impl VideoPump {
         self.stats.bytes.fetch_add(frame.data.len() as u64, Ordering::Relaxed);
         self.heartbeat();
 
-        // A piece of an AU whose predecessor went missing has nowhere to go: the decoder holds a
-        // truncated input and only a re-anchor clears it. `Discard` says so; `Feed` carries whether
-        // the pieces so far leave this AU decodable.
-        let PartStep::Feed { partial, lost_parts } = self.parts.step(frame, self.stage.takes_partial_au()) else {
-            return;
-        };
-        // Only a completed AU is a frame — a session feeding pieces would otherwise count one
-        // picture several times, and the overlay reads this figure as pictures per second.
-        if !partial {
-            self.stats.frames.fetch_add(1, Ordering::Relaxed);
-        }
-
-        let loss = self.note_loss(frame) || lost_parts;
-        let flags = FrameFlags {
+        // Everything wire-shaped, and nothing else: whether this delivery is decodable at all,
+        // and how one AU's pieces fit together, is the stage's bookkeeping.
+        let wire = WireFrame {
+            data: &frame.data,
+            pts_ns: frame.pts_ns,
+            index: frame.frame_index,
+            part: frame.part,
             reanchor: frame.flags & u32::from(FLAG_SOF) != 0 || frame.flags & USER_FLAG_RECOVERY_ANCHOR != 0,
-            loss,
-            index: u64::from(frame.frame_index),
-            partial,
+            loss: self.note_loss(frame),
         };
-        match self.stage.submit(&frame.data, frame.pts_ns, flags) {
+        match self.stage.submit(&wire) {
             SinkResult::Presented { decode_us } => {
                 if let Some(us) = decode_us {
                     self.client.report_decode_us(us);
                 }
             }
-            // Either the piece never reached the decoder (a hold swallows it, an error refused
-            // it) or it did and a keyframe has just been asked for regardless — on both paths the
-            // AU cannot be completed. Forgetting it costs the rest of one AU; keeping it would
-            // eventually feed a frame with a hole in it.
-            SinkResult::Held => self.parts.drop_open(),
+            SinkResult::Held => {}
             SinkResult::NeedKeyframe => {
-                self.parts.drop_open();
                 if let Err(e) = self.client.request_keyframe() {
                     tracing::warn!("request_keyframe: {e:#}");
                 }
